TESB416-201 Migrated picture controller to REST
