+++ web app/src/main/java/de/tinf15b4/kino/web/views/MovieView.java
@@ -22,12 +22,12 @@ import com.vaadin.ui.Link;
 import com.vaadin.ui.Panel;
 import com.vaadin.ui.VerticalLayout;
 
+import de.tinf15b4.kino.api.rest.PictureController;
 import de.tinf15b4.kino.api.rest.RestResponse;
 import de.tinf15b4.kino.data.movies.Movie;
 import de.tinf15b4.kino.data.playlists.Playlist;
 import de.tinf15b4.kino.data.ratedmovies.RatedMovie;
 import de.tinf15b4.kino.data.users.UserBean;
-import de.tinf15b4.kino.web.controllers.PictureController;
 
 @SpringView(name = MovieView.VIEW_NAME)
 public class MovieView extends VerticalLayout implements View {
